myMxShop/requirements.txt
django
pillow                  #图片支持
mysqlclient
djangorestframework-jwt
django-import-export
djangorestframework
django-bootstrap-form
markdown
django-filter
django-crispy-forms
django-reversion
django-formtools
future
httplib2
six
pymysql
coreapi                  #drf的文档支持
django-guardian          #drf对象级别的权限支持